(move) manifest plugin to its own plugin module
