+++ core/exports/plugins.ts
@@ -1,2 +1,3 @@
 export { pluginStripTypes } from "../src/plugins/strip-types.ts";
 export { pluginRadish } from "../src/plugins/plugins.ts";
+export { pluginManifest, updateManifest } from "../src/plugins/manifest.ts";
